backend/requirements.txt
python-multipart
fastapi
uvicorn
PyPDF2
nltk
gtts
farm-haystack[all]
faiss-cpu
